fix js build by moving the concurrency test to the jvm folder
